+++ lib/src/test/kotlin/me/xizzhu/android/logger/LoggerTest.kt
@@ -16,18 +16,14 @@
 
 package me.xizzhu.android.logger
 
-import org.mockito.Mock
-import org.mockito.Mockito.*
-import org.mockito.MockitoAnnotations
-import kotlin.test.AfterTest
+import io.mockk.every
+import io.mockk.mockk
+import io.mockk.verify
 import kotlin.test.BeforeTest
 import kotlin.test.Test
 
 class LoggerTest {
-    @Mock
     private lateinit var logger: Logger
-
-    private var mockCloseable: AutoCloseable? = null
 
     private val tag = "TAG"
     private val msg = "msg"
@@ -35,66 +31,39 @@ class LoggerTest {
 
     @BeforeTest
     fun setup() {
-        mockCloseable = MockitoAnnotations.openMocks(this)
-    }
-
-    @AfterTest
-    fun tearDown() {
-        mockCloseable?.close()
+        logger = mockk()
+        every { logger.log(any(), any(), any()) } returns Unit
+        every { logger.log(any(), any(), any(), any()) } returns Unit
     }
 
     @Test
     fun testNoLogger() {
         Log.v(tag, msg)
-        verify(logger, never()).log(Log.VERBOSE, tag, msg)
-        verify(logger, never()).log(Log.VERBOSE, tag, msg, exception)
-
         Log.d(tag, msg)
-        verify(logger, never()).log(Log.DEBUG, tag, msg)
-        verify(logger, never()).log(Log.DEBUG, tag, msg, exception)
-
         Log.i(tag, msg)
-        verify(logger, never()).log(Log.INFO, tag, msg)
-        verify(logger, never()).log(Log.INFO, tag, msg, exception)
-
         Log.w(tag, msg)
-        verify(logger, never()).log(Log.WARN, tag, msg)
-        verify(logger, never()).log(Log.WARN, tag, msg, exception)
-
         Log.e(tag, msg)
-        verify(logger, never()).log(Log.ERROR, tag, msg)
-        verify(logger, never()).log(Log.ERROR, tag, msg, exception)
-
         Log.f(tag, msg)
-        verify(logger, never()).log(Log.FATAL, tag, msg)
-        verify(logger, never()).log(Log.FATAL, tag, msg, exception)
+
+        verify(exactly = 0) {
+            logger.log(any(), any(), any())
+            logger.log(any(), any(), any(), any())
+        }
     }
 
     @Test
     fun testNoLoggerWithException() {
         Log.v(tag, msg, exception)
-        verify(logger, never()).log(Log.VERBOSE, tag, msg)
-        verify(logger, never()).log(Log.VERBOSE, tag, msg, exception)
-
         Log.d(tag, msg, exception)
-        verify(logger, never()).log(Log.DEBUG, tag, msg)
-        verify(logger, never()).log(Log.DEBUG, tag, msg, exception)
-
         Log.i(tag, msg, exception)
-        verify(logger, never()).log(Log.INFO, tag, msg)
-        verify(logger, never()).log(Log.INFO, tag, msg, exception)
-
         Log.w(tag, msg, exception)
-        verify(logger, never()).log(Log.WARN, tag, msg)
-        verify(logger, never()).log(Log.WARN, tag, msg, exception)
-
         Log.e(tag, msg, exception)
-        verify(logger, never()).log(Log.ERROR, tag, msg)
-        verify(logger, never()).log(Log.ERROR, tag, msg, exception)
-
         Log.f(tag, msg, exception)
-        verify(logger, never()).log(Log.FATAL, tag, msg)
-        verify(logger, never()).log(Log.FATAL, tag, msg, exception)
+
+        verify(exactly = 0) {
+            logger.log(any(), any(), any())
+            logger.log(any(), any(), any(), any())
+        }
     }
 
     @Test
@@ -103,28 +72,16 @@ class LoggerTest {
         Log.removeLogger(logger)
 
         Log.v(tag, msg)
-        verify(logger, never()).log(Log.VERBOSE, tag, msg)
-        verify(logger, never()).log(Log.VERBOSE, tag, msg, exception)
-
         Log.d(tag, msg)
-        verify(logger, never()).log(Log.DEBUG, tag, msg)
-        verify(logger, never()).log(Log.DEBUG, tag, msg, exception)
-
         Log.i(tag, msg)
-        verify(logger, never()).log(Log.INFO, tag, msg)
-        verify(logger, never()).log(Log.INFO, tag, msg, exception)
-
         Log.w(tag, msg)
-        verify(logger, never()).log(Log.WARN, tag, msg)
-        verify(logger, never()).log(Log.WARN, tag, msg, exception)
-
         Log.e(tag, msg)
-        verify(logger, never()).log(Log.ERROR, tag, msg)
-        verify(logger, never()).log(Log.ERROR, tag, msg, exception)
-
         Log.f(tag, msg)
-        verify(logger, never()).log(Log.FATAL, tag, msg)
-        verify(logger, never()).log(Log.FATAL, tag, msg, exception)
+
+        verify(exactly = 0) {
+            logger.log(any(), any(), any())
+            logger.log(any(), any(), any(), any())
+        }
     }
 
     @Test
@@ -133,28 +90,16 @@ class LoggerTest {
         Log.removeLogger(logger)
 
         Log.v(tag, msg, exception)
-        verify(logger, never()).log(Log.VERBOSE, tag, msg)
-        verify(logger, never()).log(Log.VERBOSE, tag, msg, exception)
-
         Log.d(tag, msg, exception)
-        verify(logger, never()).log(Log.DEBUG, tag, msg)
-        verify(logger, never()).log(Log.DEBUG, tag, msg, exception)
-
         Log.i(tag, msg, exception)
-        verify(logger, never()).log(Log.INFO, tag, msg)
-        verify(logger, never()).log(Log.INFO, tag, msg, exception)
-
         Log.w(tag, msg, exception)
-        verify(logger, never()).log(Log.WARN, tag, msg)
-        verify(logger, never()).log(Log.WARN, tag, msg, exception)
-
         Log.e(tag, msg, exception)
-        verify(logger, never()).log(Log.ERROR, tag, msg)
-        verify(logger, never()).log(Log.ERROR, tag, msg, exception)
-
         Log.f(tag, msg, exception)
-        verify(logger, never()).log(Log.FATAL, tag, msg)
-        verify(logger, never()).log(Log.FATAL, tag, msg, exception)
+
+        verify(exactly = 0) {
+            logger.log(any(), any(), any())
+            logger.log(any(), any(), any(), any())
+        }
     }
 
     @Test
@@ -163,28 +108,38 @@ class LoggerTest {
         Log.addLogger(logger)
 
         Log.v(tag, msg)
-        verify(logger, times(1)).log(Log.VERBOSE, tag, msg)
-        verify(logger, never()).log(Log.VERBOSE, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.VERBOSE, tag, msg)
+        }
 
         Log.d(tag, msg)
-        verify(logger, times(1)).log(Log.DEBUG, tag, msg)
-        verify(logger, never()).log(Log.DEBUG, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.DEBUG, tag, msg)
+        }
 
         Log.i(tag, msg)
-        verify(logger, times(1)).log(Log.INFO, tag, msg)
-        verify(logger, never()).log(Log.INFO, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.INFO, tag, msg)
+        }
 
         Log.w(tag, msg)
-        verify(logger, times(1)).log(Log.WARN, tag, msg)
-        verify(logger, never()).log(Log.WARN, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.WARN, tag, msg)
+        }
 
         Log.e(tag, msg)
-        verify(logger, times(1)).log(Log.ERROR, tag, msg)
-        verify(logger, never()).log(Log.ERROR, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.ERROR, tag, msg)
+        }
 
         Log.f(tag, msg)
-        verify(logger, times(1)).log(Log.FATAL, tag, msg)
-        verify(logger, never()).log(Log.FATAL, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.FATAL, tag, msg)
+        }
+
+        verify(exactly = 0) {
+            logger.log(any(), any(), any(), any())
+        }
 
         Log.removeLogger(logger)
     }
@@ -195,28 +150,38 @@ class LoggerTest {
         Log.addLogger(logger)
 
         Log.v(tag, msg, exception)
-        verify(logger, never()).log(Log.VERBOSE, tag, msg)
-        verify(logger, times(1)).log(Log.VERBOSE, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.VERBOSE, tag, msg, exception)
+        }
 
         Log.d(tag, msg, exception)
-        verify(logger, never()).log(Log.DEBUG, tag, msg)
-        verify(logger, times(1)).log(Log.DEBUG, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.DEBUG, tag, msg, exception)
+        }
 
         Log.i(tag, msg, exception)
-        verify(logger, never()).log(Log.INFO, tag, msg)
-        verify(logger, times(1)).log(Log.INFO, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.INFO, tag, msg, exception)
+        }
 
         Log.w(tag, msg, exception)
-        verify(logger, never()).log(Log.WARN, tag, msg)
-        verify(logger, times(1)).log(Log.WARN, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.WARN, tag, msg, exception)
+        }
 
         Log.e(tag, msg, exception)
-        verify(logger, never()).log(Log.ERROR, tag, msg)
-        verify(logger, times(1)).log(Log.ERROR, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.ERROR, tag, msg, exception)
+        }
 
         Log.f(tag, msg, exception)
-        verify(logger, never()).log(Log.FATAL, tag, msg)
-        verify(logger, times(1)).log(Log.FATAL, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.FATAL, tag, msg, exception)
+        }
+
+        verify(exactly = 0) {
+            logger.log(any(), any(), any())
+        }
 
         Log.removeLogger(logger)
     }
@@ -226,28 +191,38 @@ class LoggerTest {
         Log.addLogger(logger)
 
         Log.v(tag, msg)
-        verify(logger, times(1)).log(Log.VERBOSE, tag, msg)
-        verify(logger, never()).log(Log.VERBOSE, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.VERBOSE, tag, msg)
+        }
 
         Log.d(tag, msg)
-        verify(logger, times(1)).log(Log.DEBUG, tag, msg)
-        verify(logger, never()).log(Log.DEBUG, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.DEBUG, tag, msg)
+        }
 
         Log.i(tag, msg)
-        verify(logger, times(1)).log(Log.INFO, tag, msg)
-        verify(logger, never()).log(Log.INFO, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.INFO, tag, msg)
+        }
 
         Log.w(tag, msg)
-        verify(logger, times(1)).log(Log.WARN, tag, msg)
-        verify(logger, never()).log(Log.WARN, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.WARN, tag, msg)
+        }
 
         Log.e(tag, msg)
-        verify(logger, times(1)).log(Log.ERROR, tag, msg)
-        verify(logger, never()).log(Log.ERROR, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.ERROR, tag, msg)
+        }
 
         Log.f(tag, msg)
-        verify(logger, times(1)).log(Log.FATAL, tag, msg)
-        verify(logger, never()).log(Log.FATAL, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.FATAL, tag, msg)
+        }
+
+        verify(exactly = 0) {
+            logger.log(any(), any(), any(), any())
+        }
 
         Log.removeLogger(logger)
     }
@@ -257,28 +232,38 @@ class LoggerTest {
         Log.addLogger(logger)
 
         Log.v(tag, msg, exception)
-        verify(logger, never()).log(Log.VERBOSE, tag, msg)
-        verify(logger, times(1)).log(Log.VERBOSE, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.VERBOSE, tag, msg, exception)
+        }
 
         Log.d(tag, msg, exception)
-        verify(logger, never()).log(Log.DEBUG, tag, msg)
-        verify(logger, times(1)).log(Log.DEBUG, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.DEBUG, tag, msg, exception)
+        }
 
         Log.i(tag, msg, exception)
-        verify(logger, never()).log(Log.INFO, tag, msg)
-        verify(logger, times(1)).log(Log.INFO, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.INFO, tag, msg, exception)
+        }
 
         Log.w(tag, msg, exception)
-        verify(logger, never()).log(Log.WARN, tag, msg)
-        verify(logger, times(1)).log(Log.WARN, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.WARN, tag, msg, exception)
+        }
 
         Log.e(tag, msg, exception)
-        verify(logger, never()).log(Log.ERROR, tag, msg)
-        verify(logger, times(1)).log(Log.ERROR, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.ERROR, tag, msg, exception)
+        }
 
         Log.f(tag, msg, exception)
-        verify(logger, never()).log(Log.FATAL, tag, msg)
-        verify(logger, times(1)).log(Log.FATAL, tag, msg, exception)
+        verify(exactly = 1) {
+            logger.log(Log.FATAL, tag, msg, exception)
+        }
+
+        verify(exactly = 0) {
+            logger.log(any(), any(), any())
+        }
 
         Log.removeLogger(logger)
     }